Do not remove ignored files
In the event that a file is committed in as a project specific override, we do not want this file to be cleaned up. The project has taken the responsibility of managing that file and its versioning

If we do not have this then the files in the ignored list get removed and do not get replaced.